--- src/main/java/com/studentorganizer/controllers/StudentController.java
@@ -1,59 +1,0 @@
-package com.studentorganizer.controllers;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-
-import com.studentorganizer.models.Student;
-import com.studentorganizer.models.DTO.StudentDTO;
-import com.studentorganizer.services.StudentService;
-
-
-@Controller
-public class StudentController {
-    
-    @Autowired
-    private StudentService studentService;
-
-
-    @RequestMapping("/students")
-    public String showAll(Model model){
-        model.addAttribute("students", studentService.findAll());
-        return "students";
-    }
- 
-    @GetMapping("/new-student")
-    public String showNewStudentform(Model model, StudentDTO student){
-        model.addAttribute("student", student);
-        return "new-student";
-    }
-
-    @PostMapping("/new-student")
-    public String addStudent(Model model, StudentDTO student){
-        model.addAttribute("student", student);
-        studentService.addStudent(student);
-        return "redirect:/students";
-    }
-
-    @GetMapping("/update-student/{id}")
-    public String showUpdateStudent(Model model, Student student, @PathVariable String id){
-        model.addAttribute("student", studentService.getStudentById(id));
-        return "update-student" ;
-    }
-
-   @PostMapping("/update-student")
-    public String updateStudent(Model model, Student student){
-        model.addAttribute("student", student);  
-        studentService.updateStudent(student);
-        return "redirect:/students";
-    }
-
-    @GetMapping("/delete-student/{id}")
-    public String deleteStudent(Model model, @PathVariable String id, StudentDTO student){
-        model.addAttribute("student", student);
-        studentService.deleteStudent(id);
-        return "redirect:/students";
-    }
-
-}
