--- src/main/java/com/studentorganizer/models/abs/AbstractStudent.java
@@ -1,10 +1,0 @@
-package com.studentorganizer.models.abs;
-
-public abstract class AbstractStudent {
-    private String neptune;
-    private String firstName;
-    private String lastName;
-    private String major;
-    private String educationType;
-    private String uuid;
-}
