+++ src/main/java/dev/exonbyte/student_organizer/StudentOrganizerApplication.java
@@ -1,0 +1,13 @@
+package dev.exonbyte.student_organizer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StudentOrganizerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(StudentOrganizerApplication.class, args);
+	}
+
+}
