+++ src/main/java/dev/exonbyte/student_organizer/student/AbstractStudent.java
@@ -1,0 +1,10 @@
+package dev.exonbyte.student_organizer.student;
+
+public abstract class AbstractStudent {
+    private String neptune;
+    private String firstName;
+    private String lastName;
+    private String major;
+    private String educationType;
+    private String uuid;
+}
