--- src/main/resources/python/deploy.py
@@ -1,1 +1,0 @@
-print("hello World")
