--- src/test/java/com/studentorganizer/start/StartApplicationTests.java
@@ -1,13 +1,0 @@
-package com.studentorganizer.start;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StartApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
