+++ src/test/java/dev/exonbyte/student_organizer/StudentOrganizerApplicationTests.java
@@ -1,0 +1,13 @@
+package dev.exonbyte.student_organizer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class StudentOrganizerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
